+++ Meadowlark-js/packages/meadowlark-core/src/handler/FrontendFacade.ts
@@ -22,7 +22,7 @@ import { ensurePluginsLoaded, getDocumentStore } from '../plugin/PluginLoader';
 import { queryValidation } from '../middleware/ValidateQueryMiddleware';
 import { documentInfoExtraction } from '../middleware/ExtractDocumentInfoMiddleware';
 import { metaEdModelFinding } from '../middleware/FindMetaEdModelMiddleware';
-import { anonymizeAndLogRequestBody } from '../middleware/LogAnonymizedRequestMiddleware';
+import { anonymizeAndLogRequestBody } from '../middleware/LogRequestMiddleware';
 import { logTheResponse } from '../middleware/ResponseLoggingMiddleware';
 
 type MiddlewareStack = (model: MiddlewareModel) => Promise<MiddlewareModel>;
